+++ packages/google_maps_flutter/example/lib/map_ui.dart
@@ -38,8 +38,8 @@ class MapUiBodyState extends State<MapUiBody> {
     zoom: 11.0,
   );
 
-  GoogleMapController mapController;
   CameraPosition _position = _kInitialPosition;
+  bool _isMapCreated = false;
   bool _isMoving = false;
   bool _compassEnabled = true;
   CameraTargetBounds _cameraTargetBounds = CameraTargetBounds.unbounded;
@@ -66,20 +66,8 @@ class MapUiBodyState extends State<MapUiBody> {
     });
   }
 
-  void _onMapChanged() {
-    setState(() {
-      _extractMapInfo();
-    });
-  }
-
-  void _extractMapInfo() {
-    _position = mapController.cameraPosition;
-    _isMoving = mapController.isCameraMoving;
-  }
-
   @override
   void dispose() {
-    mapController.removeListener(_onMapChanged);
     super.dispose();
   }
 
@@ -199,7 +187,6 @@ class MapUiBodyState extends State<MapUiBody> {
     final GoogleMap googleMap = GoogleMap(
       onMapCreated: onMapCreated,
       initialCameraPosition: _kInitialPosition,
-      trackCameraPosition: true,
       compassEnabled: _compassEnabled,
       cameraTargetBounds: _cameraTargetBounds,
       minMaxZoomPreference: _minMaxZoomPreference,
@@ -210,6 +197,7 @@ class MapUiBodyState extends State<MapUiBody> {
       zoomGesturesEnabled: _zoomGesturesEnabled,
       myLocationEnabled: _myLocationEnabled,
       mapStyle: _mapStyle,
+      onCameraMove: _updateCameraPosition,
     );
 
     final List<Widget> columnChildren = <Widget>[
@@ -225,7 +213,7 @@ class MapUiBodyState extends State<MapUiBody> {
       ),
     ];
 
-    if (mapController != null) {
+    if (_isMapCreated) {
       columnChildren.add(
         Expanded(
           child: ListView(
@@ -258,10 +246,15 @@ class MapUiBodyState extends State<MapUiBody> {
     );
   }
 
+  void _updateCameraPosition(CameraPosition position) {
+    setState(() {
+      _position = position;
+    });
+  }
+
   void onMapCreated(GoogleMapController controller) {
-    mapController = controller;
-    mapController.addListener(_onMapChanged);
-    _extractMapInfo();
-    setState(() {});
+    setState(() {
+      _isMapCreated = true;
+    });
   }
 }