+++ packages/in_app_purchase/test/store_kit_wrappers/sk_product_test.dart
@@ -5,54 +5,18 @@
 import 'package:test/test.dart';
 import 'package:in_app_purchase/src/store_kit_wrappers/sk_product_wrapper.dart';
 import 'package:in_app_purchase/src/in_app_purchase_connection/product_details.dart';
+import 'package:in_app_purchase/store_kit_wrappers.dart';
+import 'sk_test_stub_objects.dart';
 
 void main() {
-  final Map<String, dynamic> localeMap = <String, dynamic>{
-    'currencySymbol': '\$'
-  };
-  final Map<String, dynamic> subMap = <String, dynamic>{
-    'numberOfUnits': 1,
-    'unit': 2
-  };
-  final Map<String, dynamic> discountMap = <String, dynamic>{
-    'price': '1.0',
-    'priceLocale': localeMap,
-    'numberOfPeriods': 1,
-    'paymentMode': 2,
-    'subscriptionPeriod': subMap,
-  };
-  final Map<String, dynamic> productMap = <String, dynamic>{
-    'productIdentifier': 'id',
-    'localizedTitle': 'title',
-    'localizedDescription': 'description',
-    'priceLocale': localeMap,
-    'downloadContentVersion': 'version',
-    'subscriptionGroupIdentifier': 'com.group',
-    'price': '1.0',
-    'downloadable': true,
-    'downloadContentLengths': <int>[1, 2],
-    'subscriptionPeriod': subMap,
-    'introductoryPrice': discountMap,
-  };
-
-  final Map<String, List<dynamic>> productResponseMap = <String, List<dynamic>>{
-    'products': <Map<String, dynamic>>[productMap],
-    'invalidProductIdentifiers': <String>['123'],
-  };
-
-  group('product request wrapper test', () {
-    void testMatchLocale(
-        PriceLocaleWrapper wrapper, Map<String, dynamic> localeMap) {
-      expect(wrapper.currencySymbol, localeMap['currencySymbol']);
-    }
-
+  group('product related object wrapper test', () {
     test(
         'SKProductSubscriptionPeriodWrapper should have property values consistent with map',
         () {
       final SKProductSubscriptionPeriodWrapper wrapper =
-          SKProductSubscriptionPeriodWrapper.fromJson(subMap);
-      expect(wrapper.numberOfUnits, subMap['numberOfUnits']);
-      expect(wrapper.unit, SubscriptionPeriodUnit.values[subMap['unit']]);
+          SKProductSubscriptionPeriodWrapper.fromJson(
+              buildSubscriptionPeriodMap(dummySubscription));
+      expect(wrapper, equals(dummySubscription));
     });
 
     test(
@@ -68,18 +32,8 @@ void main() {
         'SKProductDiscountWrapper should have property values consistent with map',
         () {
       final SKProductDiscountWrapper wrapper =
-          SKProductDiscountWrapper.fromJson(discountMap);
-      expect(wrapper.price, discountMap['price']);
-      testMatchLocale(wrapper.priceLocale, discountMap['priceLocale']);
-      expect(wrapper.numberOfPeriods, discountMap['numberOfPeriods']);
-      expect(wrapper.paymentMode,
-          ProductDiscountPaymentMode.values[discountMap['paymentMode']]);
-      expect(
-          wrapper.subscriptionPeriod.unit,
-          SubscriptionPeriodUnit
-              .values[discountMap['subscriptionPeriod']['unit']]);
-      expect(wrapper.subscriptionPeriod.numberOfUnits,
-          discountMap['subscriptionPeriod']['numberOfUnits']);
+          SKProductDiscountWrapper.fromJson(buildDiscountMap(dummyDiscount));
+      expect(wrapper, equals(dummyDiscount));
     });
 
     test(
@@ -94,49 +48,11 @@ void main() {
       expect(wrapper.subscriptionPeriod, null);
     });
 
-    void testMatchingProductMap(
-        SKProductWrapper wrapper, Map<String, dynamic> productMap) {
-      expect(wrapper.productIdentifier, productMap['productIdentifier']);
-      expect(wrapper.localizedTitle, productMap['localizedTitle']);
-      testMatchLocale(wrapper.priceLocale, productMap['priceLocale']);
-      expect(wrapper.localizedDescription, productMap['localizedDescription']);
-      expect(
-          wrapper.downloadContentVersion, productMap['downloadContentVersion']);
-      expect(wrapper.subscriptionGroupIdentifier,
-          productMap['subscriptionGroupIdentifier']);
-      expect(wrapper.price, productMap['price']);
-      expect(wrapper.downloadable, productMap['downloadable']);
-      expect(
-          wrapper.downloadContentLengths, productMap['downloadContentLengths']);
-      expect(wrapper.introductoryPrice.price,
-          productMap['introductoryPrice']['price']);
-      expect(wrapper.introductoryPrice.numberOfPeriods,
-          productMap['introductoryPrice']['numberOfPeriods']);
-      expect(
-          wrapper.introductoryPrice.paymentMode,
-          ProductDiscountPaymentMode
-              .values[productMap['introductoryPrice']['paymentMode']]);
-      expect(
-          wrapper.introductoryPrice.subscriptionPeriod.unit,
-          SubscriptionPeriodUnit.values[productMap['introductoryPrice']
-              ['subscriptionPeriod']['unit']]);
-      expect(
-          wrapper.introductoryPrice.subscriptionPeriod.numberOfUnits,
-          productMap['introductoryPrice']['subscriptionPeriod']
-              ['numberOfUnits']);
-      expect(
-          wrapper.subscriptionPeriod.unit,
-          SubscriptionPeriodUnit
-              .values[productMap['subscriptionPeriod']['unit']]);
-      expect(wrapper.subscriptionPeriod.numberOfUnits,
-          productMap['subscriptionPeriod']['numberOfUnits']);
-      expect(wrapper.price, discountMap['price']);
-    }
-
     test('SKProductWrapper should have property values consistent with map',
         () {
-      final SKProductWrapper wrapper = SKProductWrapper.fromJson(productMap);
-      testMatchingProductMap(wrapper, productMap);
+      final SKProductWrapper wrapper =
+          SKProductWrapper.fromJson(buildProductMap(dummyProductWrapper));
+      expect(wrapper, equals(dummyProductWrapper));
     });
 
     test('SKProductWrapper should have properties to be null if map is empty',
@@ -155,7 +71,8 @@ void main() {
     });
 
     test('toProductDetails() should return correct Product object', () {
-      final SKProductWrapper wrapper = SKProductWrapper.fromJson(productMap);
+      final SKProductWrapper wrapper =
+          SKProductWrapper.fromJson(buildProductMap(dummyProductWrapper));
       final ProductDetails product = wrapper.toProductDetails();
       expect(product.title, wrapper.localizedTitle);
       expect(product.description, wrapper.localizedDescription);
@@ -166,11 +83,9 @@ void main() {
 
     test('SKProductResponse wrapper should match', () {
       final SkProductResponseWrapper wrapper =
-          SkProductResponseWrapper.fromJson(productResponseMap);
-      testMatchingProductMap(
-          wrapper.products[0], productResponseMap['products'][0]);
-      expect(wrapper.invalidProductIdentifiers,
-          productResponseMap['invalidProductIdentifiers']);
+          SkProductResponseWrapper.fromJson(
+              buildProductResponseMap(dummyProductResponseWrapper));
+      expect(wrapper, equals(dummyProductResponseWrapper));
     });
     test('SKProductResponse wrapper should default to empty list', () {
       final Map<String, List<dynamic>> productResponseMapEmptyList =
@@ -185,8 +100,48 @@ void main() {
     });
 
     test('LocaleWrapper should have property values consistent with map', () {
-      final PriceLocaleWrapper wrapper = PriceLocaleWrapper.fromJson(localeMap);
-      testMatchLocale(wrapper, localeMap);
+      final SKPriceLocaleWrapper wrapper =
+          SKPriceLocaleWrapper.fromJson(buildLocaleMap(dummyLocale));
+      expect(wrapper, equals(dummyLocale));
+    });
+  });
+
+  group('Payment queue related object tests', () {
+    test('Should construct correct SKPaymentWrapper from json', () {
+      SKPaymentWrapper payment =
+          SKPaymentWrapper.fromJson(dummyPayment.toMap());
+      expect(payment, equals(dummyPayment));
+    });
+
+    test('Should construct correct SKError from json', () {
+      SKError error = SKError.fromJson(buildErrorMap(dummyError));
+      expect(error, equals(dummyError));
+    });
+
+    test('Should construct correct SKDownloadWrapper from json', () {
+      SKDownloadWrapper download =
+          SKDownloadWrapper.fromJson(buildDownloadMap(dummyDownload));
+      expect(download, equals(dummyDownload));
+    });
+
+    test('Should construct correct SKTransactionWrapper from json', () {
+      SKPaymentTransactionWrapper transaction =
+          SKPaymentTransactionWrapper.fromJson(
+              buildTransactionMap(dummyTransaction));
+      expect(transaction, equals(dummyTransaction));
+    });
+
+    test('Should generate correct map of the payment object', () {
+      Map map = dummyPayment.toMap();
+      expect(map['productIdentifier'], dummyPayment.productIdentifier);
+      expect(map['applicationUsername'], dummyPayment.applicationUsername);
+
+      expect(map['requestData'], dummyPayment.requestData);
+
+      expect(map['quantity'], dummyPayment.quantity);
+
+      expect(map['simulatesAskToBuyInSandbox'],
+          dummyPayment.simulatesAskToBuyInSandbox);
     });
   });
 }