+++ packages/google_maps_flutter/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.4.0
+
+* Change events are call backs on GoogleMap widget.
+* GoogleMapController no longer handles change events.
+* trackCameraPosition is inferred from GoogleMap.onCameraMove being set.
+
 ## 0.3.0+3
 
 * Update Android play-services-maps to 16.1.0